renaming helloController - entry point of our applicaiton.
