+++ poster-spring-boot/src/test/java/org/pti/poster/boot/HelloControllerIT.java
@@ -1,4 +1,4 @@
-package hello;
+package org.pti.poster.boot;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
